pyproject.toml
[project]
name = "graph-code"
version = "0.0.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "langgraph>=0.2.66",
    "langchain>=0.3.13",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.2.14",
    "pydantic-settings>=2.0.0",
    "pymgclient>=1.4.0",
    "python-dotenv>=1.1.0",
    "toml>=0.10.2",
    "tree-sitter-python>=0.23.6",
    "tree-sitter==0.25.0",
    "watchdog>=6.0.0",
    "typer>=0.12.5",
    "rich>=13.7.1",
    "prompt-toolkit>=3.0.0",
    "diff-match-patch>=20241021",
    "click>=8.0.0",
]

[tool.setuptools]
packages = ["codebase_rag"]

[project.optional-dependencies]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
]
dev = [
    "ruff>=0.5.5",
    "mypy>=1.11.0",
    "pre-commit>=4.2.0",
]

treesitter-full = [
    "tree-sitter-python>=0.23.6",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-rust>=0.24.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-scala>=0.24.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-cpp>=0.23.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_defs = true

[dependency-groups]
dev = [
    "pyinstaller>=6.14.1",
    "types-toml>=0.10.8.20240310",
]
